talks_processing_requirements.txt
pyyaml
pydub